Moved vertical profile chart to a new file
