+++ SubmersionTimeSeriesChart.swift
@@ -1,0 +1,18 @@
+//
+//  SubmersionTimeSeriesChart.swift
+//  BoB
+//
+//  Created by Kim Martini on 6/13/25.
+//
+
+import SwiftUI
+
+struct SubmersionTimeSeriesChart: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    SubmersionTimeSeriesChart()
+}
